requirements.txt
aioredis==1.2.0
asgi-redis==1.4.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.1
djangorestframework==3.9.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
msgpack-python==0.5.6
mysqlclient==1.4.2.post1
PyHamcrest==1.9.0
python-decouple==3.1
pytz==2019.1
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
whitenoise==4.1.2
zope.interface==4.6.0
